Move over NumPyConverter code, small refactorings. Now very broken
